Refactored FileSystemPicker
